vortex/backend-analyzer/requirements.txt
fastapi
uvicorn
google-adk>=0.7.0
litellm
python-dotenv
pydantic